+++ agentlib_mpc/__init__.py
@@ -4,4 +4,4 @@ and data-driven modeling."""
 from .modules import MODULE_TYPES
 from .models import MODEL_TYPES
 
-__version__ = "0.6.7"
+__version__ = "0.6.8"